dev-InterProject/InterProject/jobapps/user/serializers.py
from rest_framework import serializers

from jobapps.user.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
        depth = 2  # 深度



