dev-InterProject/InterProject/requirements-know.txt
# Graphapps知识图谱功能依赖包
# 专业知识图谱自动生成系统所需的Python包

# ==================== 核心框架 ====================
# Django Web框架
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# ==================== 图数据库 ====================
# Neo4j图数据库驱动
neo4j>=5.0.0,<6.0.0

# ==================== 数据处理 ====================
# 数据分析和处理
pandas>=1.5.0,<2.0.0
numpy>=1.24.0,<2.0.0

# JSON数据处理和验证
jsonschema>=4.19.0

# ==================== 自然语言处理 ====================
# 中文分词
jieba>=0.42.1

# 文本相似度计算
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0  # fuzzywuzzy的加速依赖

# ==================== 网络请求 ====================
# HTTP请求库
requests>=2.31.0
urllib3>=1.26.0

# 异步HTTP请求
aiohttp>=3.8.0,<4.0.0

# ==================== Web爬虫和解析 ====================
# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0  # BeautifulSoup的XML解析器

# ==================== 环境和配置管理 ====================
# 环境变量管理
python-dotenv>=1.0.0

# 配置管理和数据验证
pydantic>=2.3.0,<3.0.0

# ==================== 时间和日期处理 ====================
# 日期时间处理
python-dateutil>=2.8.0
pytz>=2023.3

# ==================== 系统和性能监控 ====================
# 系统信息获取
psutil>=5.9.0

# 进度条显示
tqdm>=4.66.0

# ==================== 大数据处理（可选） ====================
# Apache Spark（用于大规模数据处理）
# pyspark>=3.4.0  # 取消注释以启用Spark功能

# ==================== Web应用扩展 ====================
# Flask（用于独立Web应用模式）
Flask>=2.3.0
Flask-CORS>=4.0.0

# Django跨域处理
django-cors-headers>=4.2.0

# ==================== 开发和测试工具（可选） ====================
# 代码格式化
# black>=23.0.0
# isort>=5.12.0

# 代码检查
# flake8>=6.0.0
# pylint>=2.17.0

# 测试框架
# pytest>=7.4.0
# pytest-django>=4.5.0
# pytest-asyncio>=0.21.0

# ==================== 生产环境部署（可选） ====================
# WSGI服务器
# gunicorn>=21.2.0

# 静态文件服务
# whitenoise>=6.5.0

# 数据库适配器（如果需要其他数据库）
# mysqlclient>=2.2.0  # MySQL
# psycopg2-binary>=2.9.0  # PostgreSQL

# ==================== 机器学习和AI（扩展功能） ====================
# 如果需要更高级的NLP功能，可以添加：
# transformers>=4.30.0
# torch>=2.0.0
# scikit-learn>=1.3.0
# nltk>=3.8.0
# spacy>=3.6.0

# ==================== 图形和可视化（可选） ====================
# 如果需要生成图谱可视化图片
# matplotlib>=3.7.0
# networkx>=3.1.0
# plotly>=5.15.0

# ==================== 缓存和存储（可选） ====================
# Redis缓存
# redis>=4.6.0
# django-redis>=5.3.0

# ==================== 安全和认证（可选） ====================
# JWT认证
# PyJWT>=2.8.0
# django-rest-framework-simplejwt>=5.2.0

# ==================== 日志和监控（可选） ====================
# 结构化日志
# structlog>=23.1.0

# 性能监控
# django-debug-toolbar>=4.1.0

# ==================== 版本说明 ====================
# Python版本要求: >=3.8,<4.0
# 推荐使用Python 3.9或3.10以获得最佳性能和兼容性

# ==================== 安装说明 ====================
# 基础安装（仅核心功能）：
# pip install -r requirements.txt

# 完整安装（包含所有可选功能）：
# 取消注释相应的包并运行：
# pip install -r requirements.txt

# 开发环境安装：
# pip install -r requirements.txt
# 然后取消注释开发工具部分的包

# ==================== 特殊说明 ====================
# 1. Neo4j数据库需要单独安装和配置
# 2. 某些包可能需要系统级依赖，请参考各包的官方文档
# 3. 在Windows系统上，某些包可能需要Visual C++编译器
# 4. 建议使用虚拟环境来管理依赖
